+++ insts/std/STRUM/VSTART1.h
@@ -20,6 +20,6 @@ class VSTART1 : public Instrument {
 public:
 	VSTART1();
 	virtual ~VSTART1();
-	int init(float*, short);
+	int init(float*, int);
 	int run();
 	};
